requirement.txt
Django==4.2.4
djangorestframework==3.14.0
mysqlclient==2.2.0
django-mysql==4.11.0 
django-cors-headers==4.2.0
PyJWT==2.8.0
Faker==19.2.0
django-redis==5.3.0
stripe==5.5.0
Pillow==10.0.0
django-filter==23.3
twilio==8.9.0 
channels==4.0.0
